split `build_scalar` module from `function/mod.rs`
